add support for the identity hash function
With integer code 0x00 as per multicodec specs, thus adjust the range of
application-specific hash functions too.  As discussed in #6, identity
multihashes cannot be truncated, and verification against a truncated data
hash is forbidden too (as finding collisions is trivial).  This is in some
manner similar to what is allowed by go-multihash.

Resolves #6.